requirements.txt
python-telegram-bot==13.15
APScheduler==3.10.4
pytz==2024.1